Laga aðeins til i controllerunum (skipting sena)
Lagaði aðeins til í controllerunum varðandi skiptingu á senum
