+++ src/main/java/hi/verkefni/vidmot/SearchController.java
@@ -24,23 +24,7 @@ public class SearchController {
         stage.show();
     }
 
-    public void switchToSceneSearch(ActionEvent actionEvent) throws IOException {
-        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("Search.fxml")));
-        stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
-        scene = new Scene((root));
-        stage.setTitle("Flight Search");
-        stage.setScene(scene);
-        stage.show();
-    }
-
-    public void switchToSceneRegistration(ActionEvent actionEvent) throws IOException {
-        root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("registration.fxml")));
-        stage = (Stage) ((Node) actionEvent.getSource()).getScene().getWindow();
-        scene = new Scene((root));
-        stage.setTitle("Flight Registration");
-        stage.setScene(scene);
-        stage.show();
-    }
+    
 }
