+++ src/main/resources/hi/verkefni/vidmot/completed.fxml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.2"
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="hi.verkefni.vidmot.SceneController">
+    <children>
+        <Text layoutX="159.0" layoutY="126.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Bókun staðfest!">
+            <font>
+                <Font name="System Bold Italic" size="43.0"/>
+            </font>
+        </Text>
+        <Text layoutX="202.0" layoutY="169.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Booking confirmed!">
+            <font>
+                <Font size="24.0"/>
+            </font>
+        </Text>
+    </children>
+</AnchorPane>
